# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UPcar - ride sharing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List driver profiles
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, denied]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverProfile"
    post:
      operationId: createDriverProfile
      tags: [drivers]
      summary: Submit driver registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverProfileInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── AUTH ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, phone, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        totalRides:
          type: integer
        accountStatus:
          type: string
          enum: [pending, approved, denied]
        createdAt:
          type: string
          format: date-time
        driverProfile:
          $ref: "#/components/schemas/DriverProfile"
...
    DriverProfile:
      type: object
      required: [id, userId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, denied]
        cpf:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        cnhNumber:
          type: ["string", "null"]
        cnhCategory:
          type: ["string", "null"]
        cnhExpiry:
          type: ["string", "null"]
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleYear:
          type: ["integer", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehicleType:
          type: ["string", "null"]
          enum: [sedan, suv, hatch, pickup, moto, van, null]
        photoUrl:
          type: ["string", "null"]
        cnhPhotoUrl:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        criminalRecordUrl:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    DriverProfileInput:
      type: object
      required: [cpf, birthDate, address, city, state, cnhNumber, cnhCategory, cnhExpiry, vehicleMake, vehicleModel, vehicleYear, vehicleColor, vehiclePlate, vehicleType]
      properties:
        cpf:
          type: string
        birthDate:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        cnhNumber:
          type: string
        cnhCategory:
          type: string
        cnhExpiry:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehicleColor:
          type: string
        vehiclePlate:
          type: string
        vehicleType:
          type: string
          enum: [sedan, suv, hatch, pickup, moto, van]
        photoUrl:
          type: ["string", "null"]
        cnhPhotoUrl:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        criminalRecordUrl:
          type: ["string", "null"]
...